+++ src/errors.h
@@ -1,9 +1,9 @@
 #ifndef ERRORS_H
 #define ERRORS_H
 
-long double uniform_rand(void);
+double uniform_rand(void);
 double norm_rand(double mean, double std);
-void calc_statistics(long double * vals, int length, long double * mean, long double * std);
+void calc_statistics(double * vals, int length, double * mean, double * std);
 int calc_errors(struct Model *m, int residue);
